requirements.txt
pyside6
pandas
thefuzz
beautifulsoup4
openpyxl
pyyaml
openai
python-docx
pymupdf
pytest